Add aria-disabled and aria-busy props to disnbaled and busy buttons
When the submit button has the class 'disabled-button' apply the attribute `aria-disabled="true"`.

When the submit button has the class 'loading-button' apply the attribute `aria-busy="true"`.

These attributes and their values should inform the screen-reader users that the button is not interactive and loading respectively.

When the button is not disabled, the `aria-disabled` is removed and same for when the button is not loading, it should not have `aria-busy`.